+++ web/modules/custom/hoeringsportal_base_fixtures/src/Fixture/TermDepartmentFixture.php
@@ -1,0 +1,26 @@
+<?php
+
+namespace Drupal\hoeringsportal_base_fixtures\Fixture;
+
+/**
+ * Tag term fixture.
+ *
+ * @package Drupal\hoeringsportal_base_fixtures\Fixture
+ */
+class TermDepartmentFixture extends TaxonomyTermFixture {
+  /**
+   * {@inheritdoc}
+   */
+  protected static $vocabularyId = 'department';
+
+  /**
+   * Constructor.
+   */
+  public function __construct() {
+    if (empty(static::$terms)) {
+      static::$terms = array_map(static fn ($i) => 'Department ' . $i, range(1, 3));
+    }
+    parent::__construct();
+  }
+
+}